Removed commented code snippets from game_script.js. Removed const targetWords and const targetWord that was getting random value from targetWords in game_script.js. Renamed game-routes-SEANC_WIP.js to game-routes.js and updated associated path in /controllers/api/index.js.
